+++ Sources/WingedSwift/WingedSwift.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 public enum WingedSwift: String {
-    case VERSION = "1.3.3"
+    case VERSION = "1.4.0"
     case AUTHOR = "Michel Anderson Lutz Teixeira"
     case CONTACT = "https://micheltlutz.me"
 }